Add the last exercices of the hours 4

+++ Learn_C/Ex4/Ex4H4.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 int main()
 {
-	float dbl_num = 123.456;
+	double dbl_num = 123.456;
 
 	printf("this is the scientific value of the number 123.456 = %e.\n", dbl_num);
 	printf("This is the floating point value of the number 123.456 %f.\n", dbl_num);
